fix: file upload from teams and message update and delete logic
